+++ Example.hs
@@ -1,27 +1,26 @@
-{-# LANGUAGE OverloadedStrings #-}
 
 module Example where
 
-import           Control.Monad      (unless)
+import           Control.Monad (unless)
 import           Data.Aeson
-import qualified Data.Aeson.Pointer as P
+import qualified JSONPointer   as JP
 
 main :: IO ()
 main = do
     -- JSON Pointers must either be empty or start with a /.
-    pntr1 <- case P.unescape "/foo/0" of
+    pntr1 <- case JP.unescape "/foo/0" of
                  Left _     -> error "Failed to construct JSON Pointer."
                  Right pntr -> return pntr
 
     -- We can also write JSON Pointers in Haskell.
-    let pntr2 = P.Pointer [P.Token "/"]
+    let pntr2 = JP.Pointer [JP.Token "/"]
     -- When we do this we don't have to escape / or ~ characters
     -- (as ~1 and ~0 respectively) like we do in an escaped JSON
     -- Pointer string.
-    unless (P.unescape "/~1" == Right pntr2) (error "ohno!")
+    unless (JP.unescape "/~1" == Right pntr2) (error "ohno!")
 
-    print (P.resolve pntr1 document)
-    print (P.resolve pntr2 document)
+    print (JP.resolve pntr1 document)
+    print (JP.resolve pntr2 document)
 
   where
     document :: Value