+++ src/JSONPointer.hs
@@ -1,13 +1,18 @@
+{-# LANGUAGE DeriveGeneric              #-}
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
+{-# LANGUAGE OverloadedStrings          #-}
 
-module Data.Aeson.Pointer where
+module JSONPointer where
 
 import           Control.Monad       (when)
 import           Data.Aeson
-import qualified Data.HashMap.Strict as H
+import qualified Data.Hashable       as HA
+import qualified Data.HashMap.Strict as HM
+import           Data.Semigroup      (Semigroup)
 import           Data.Text           (Text)
 import qualified Data.Text           as T
 import qualified Data.Vector         as V
+import           GHC.Generics        (Generic)
 import           Test.QuickCheck
 import           Text.Read           (readMaybe)
 
@@ -16,9 +21,30 @@ import           Control.Applicative
 import           Data.Monoid
 import           Data.Traversable
 
+--------------------------------------------------
+-- * Resolution
+--------------------------------------------------
+
+data ResolutionError
+    = ObjectLookupFailed
+    | ArrayIndexInvalid
+    | ArrayElemNotFound
+    | ExpectedObjectOrArray
+    deriving (Eq, Show)
+
+resolve :: Pointer -> Value -> Either ResolutionError Value
+resolve (Pointer []) v     = Right v
+resolve (Pointer (t:ts)) v = resolveToken t v >>= resolve (Pointer ts)
+
+--------------------------------------------------
+-- * Main types and escaping
+--------------------------------------------------
+
 newtype Pointer
     = Pointer { _unPointer :: [Token] }
-    deriving (Eq, Show, Monoid, Arbitrary)
+    deriving (Eq, Show, Semigroup, Monoid, Generic, Arbitrary)
+
+instance HA.Hashable Pointer
 
 instance FromJSON Pointer where
     parseJSON = withText "JSON Pointer" $ \t ->
@@ -32,13 +58,22 @@ instance ToJSON Pointer where
 -- | We don't try to distinguish between integer tokens and string
 -- tokens since all tokens start as strings, and all tokens can
 -- be used to resolve JSON objects.
+--
+-- Since these are unescaped you can write @"/"@ and @"~"@ normally.
+-- (e.g. if you're referencing a key such as @"abc/123"@, go ahead
+-- and write that exactly.
 newtype Token
     = Token { _unToken :: Text }
-    deriving (Eq, Show)
+    deriving (Eq, Show, Generic)
+
+instance HA.Hashable Token
 
 instance Arbitrary Token where
     arbitrary = Token . T.pack <$> arbitrary
 
+-- | This escapes @"/"@ (because it's the token separator character).
+--
+-- It also escapes @"~"@ (because it's the escape character).
 escape :: Pointer -> Text
 escape (Pointer []) = ""
 escape (Pointer ts) =
@@ -47,14 +82,11 @@ escape (Pointer ts) =
     . fmap (T.replace "/" "~1" . T.replace "~" "~0" . _unToken)
     $ ts
 
--- * Unescaping
-
 data FormatError
     = InvalidFirstChar
     -- ^ JSON Pointers must either be empty or start with a @/@.
     | UnescapedTilde
     deriving (Eq, Show)
-
 
 -- | JSON Pointers must either be empty or start with a @/@. This means
 -- that if you're turning a URI Fragment into a JSON Pointer you must
@@ -77,7 +109,40 @@ unescape txt =
               Nothing  -> Left UnescapedTilde
               Just tok -> Right tok
 
--- | For internal use by 'unescape'.
+--------------------------------------------------
+-- * Wrapper Types
+--
+-- These aren't used by the rest of the library
+-- (as explained in the docs for 'Token').
+--
+-- However, they might be useful if you need to distinguish JSON Pointer
+-- tokens from plain 'Text' or 'Int' without losing information by
+-- converting to 'Token'.
+--------------------------------------------------
+
+-- | A glorified @type@ alias. If you need to do JSON Pointer operations
+-- you're looking for 'Token' instead.
+--
+-- NOTE: Unlike 'Token' this is escaped.
+newtype Key
+    = Key { _unKey :: Text }
+    deriving (Eq, Show, Generic)
+
+instance HA.Hashable Key
+
+-- | A glorified @type@ alias. If you need to do JSON Pointer operations
+-- you're looking for 'Token' instead.
+newtype Index
+    = Index { _unIndex :: Int }
+    deriving (Eq, Show, Generic)
+
+instance HA.Hashable Index
+
+--------------------------------------------------
+-- * Internals
+--------------------------------------------------
+
+-- | For internal use (by 'unescape').
 unescapeToken :: Text -> Maybe Token
 unescapeToken t
     | not (isValid t) = Nothing
@@ -93,21 +158,10 @@ unescapeToken t
     replace :: Text -> Text
     replace = T.replace "~0" "~" . T.replace "~1" "/"
 
--- * Resolution
-
-data ResolutionError
-    = ObjectLookupFailed
-    | ArrayIndexInvalid
-    | ArrayElemNotFound
-    | ExpectedObjectOrArray
-    deriving (Eq, Show)
-
-resolve :: Pointer -> Value -> Either ResolutionError Value
-resolve (Pointer []) v     = Right v
-resolve (Pointer (t:ts)) v = resolveToken t v >>= resolve (Pointer ts)
-
--- | For internal use (or specialized applications that don't want to
--- resolve an entire pointer at once).
+-- | For internal use (by 'resolve').
+--
+-- Might also be useful for specialized applications that don't
+-- want to resolve an entire pointer at once.
 resolveToken :: Token -> Value -> Either ResolutionError Value
 resolveToken tok (Array vs) =
     case readMaybe . T.unpack . _unToken $ tok of
@@ -118,7 +172,7 @@ resolveToken tok (Array vs) =
                 Nothing  -> Left ArrayElemNotFound
                 Just res -> Right res
 resolveToken tok (Object h) =
-    case H.lookup (_unToken tok) h of
+    case HM.lookup (_unToken tok) h of
         Nothing  -> Left ObjectLookupFailed
         Just res -> Right res
 resolveToken _ _ = Left ExpectedObjectOrArray